requirements.txt
typer==0.12.3
requests==2.32.3
python-dotenv==1.0.1
watchdog==4.0.1
black==24.8.0
pytest==8.3.2
pyinstaller==6.10.0
